ai-fine-tuning/requirements.txt
# AI Fine-tuning Requirements
# Python 3.8+

# OpenAI API
openai>=1.0.0
tiktoken>=0.5.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.1
argparse>=1.4.0

# Testing & Validation
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional (for advanced features)
scikit-learn>=1.3.0  # For data analysis
matplotlib>=3.7.0    # For visualization
seaborn>=0.12.0     # For advanced plots 